+++ packages/plugin-zod/src/generators/zodGenerator.test.tsx
@@ -330,7 +330,7 @@ describe('zodGenerator schema', async () => {
       output: './gen',
     })
 
-    const schemas = getSchemas({ oas })
+    const { schemas } = getSchemas({ oas })
     const name = props.path
     const schema = schemas[name] as SchemaObject
     const tree = generator.parse({ schema, name, parentName: null })